requirements.txt
accelerate==0.24.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.5
async-timeout==4.0.2
attrs==23.1.0
autoattack @ git+https://github.com/fra31/auto-attack.git@a39220048b3c9f2cca9a4d3a54604793c68eca7e
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
boltons==24.1.0
braceexpand==0.1.7
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.11.1
dill==0.3.6
docker-pycreds==0.4.0
eagerpy==0.30.0
einops==0.6.1
einops-exts==0.0.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
fonttools==4.39.3
foolbox==3.3.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
gdown==5.1.0
geotorch==0.3.0
gitdb==4.0.10
GitPython==3.1.31
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.14.1
idna==2.10
importlib_metadata==8.5.0
inflection==0.5.1
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.10.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.4
lit==16.0.3
lmdb==1.6.2
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.4
multiprocess==0.70.14
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.10.2.3
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
open-clip-torch==2.19.0
opencv-python==4.7.0.72
optree==0.15.0
overrides==7.4.0
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
propcache==0.2.0
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-fid-wrapper==0.0.4
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
referencing==0.35.1
regex==2023.5.5
requests==2.32.3
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
robustbench @ git+https://github.com/RobustBench/robustbench.git@46a91f44524133b2cd8f721ec7e73ecb63f17fc8
rpds-py==0.20.0
safetensors==0.5.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
sentencepiece==0.1.98
sentry-sdk==1.22.2
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.11.1
tensorboardX==2.0
terminado==0.17.1
threadpoolctl==3.1.0
timm==1.0.9
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.2.1
torch==2.0.1
torchdiffeq==0.2.1
torchopt==0.7.3
torchsde==0.2.5
torchvision==0.15.2
tornado==6.3.1
tqdm==4.56.1
traitlets==5.9.0
trampoline==0.1.2
transformers @ git+https://github.com/huggingface/transformers@d3cbc997a231098cca81ac27fd3028a5536abe67
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
wandb==0.15.2
wcwidth==0.2.6
webcolors==24.11.1
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.13
xvfbwrapper==0.2.9
xxhash==3.2.0
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.21.0
